requirements.txt
streamlit>=1.28.0
langchain>=0.3.26
langchain-anthropic>=0.3.7
langgraph>=0.5.0
anthropic>=0.57.0
pydantic>=2.11.0
python-dotenv>=1.1.1
pypdf2>=3.0.1
python-docx>=1.2.0
typing-extensions>=4.14.0
streamlit-pdf-viewer>=0.0.26